ci: update packer definition

+++ dev/packer/scripts/900-cleanup.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+# DigitalOcean Marketplace Image Validation Tool
+# © 2021 DigitalOcean LLC.
+# This code is licensed under Apache 2.0 license (see LICENSE.md for details)
+
+set -o errexit
+
 # Ensure /tmp exists and has the proper permissions before
 # checking for security updates
 # https://github.com/digitalocean/marketplace-partners/issues/94
@@ -8,13 +14,17 @@ if [[ ! -d /tmp ]]; then
 fi
 chmod 1777 /tmp
 
-export DEBIAN_FRONTEND=noninteractive
-apt-get -y update
-apt-get -o Dpkg::Options::="--force-confold" upgrade -q -y --force-yes
-apt-get purge droplet-agent
-rm -rf /opt/digitalocean
-apt-get -y autoremove
-apt-get -y autoclean
+if [ -n "$(command -v yum)" ]; then
+  yum update -y
+  yum clean all
+elif [ -n "$(command -v apt-get)" ]; then
+  export DEBIAN_FRONTEND=noninteractive
+  apt-get -y update
+  apt-get -o Dpkg::Options::="--force-confold" upgrade -q -y --force-yes
+  apt-get -y autoremove
+  apt-get -y autoclean
+fi
+
 rm -rf /tmp/* /var/tmp/*
 history -c
 cat /dev/null > /root/.bash_history
@@ -36,12 +46,4 @@ The secure erase will complete successfully when you see:${NC}
     dd: writing to '/zerofile': No space left on device\n
 Beginning secure erase now\n"
 
-dd if=/dev/zero of=/zerofile &
-  PID=$!
-  while [ -d /proc/$PID ]
-    do
-      printf "."
-      sleep 5
-    done
-sync; rm /zerofile; sync
-cat /dev/null > /var/log/lastlog; cat /dev/null > /var/log/wtmp
+dd if=/dev/zero of=/zerofile bs=4096 || rm /zerofile
